+++ src/Application/Game.cpp
@@ -17,7 +17,7 @@
 #include "Engine/Graphics/LightsStack.h"
 #include "Engine/Graphics/LightmapBuilder.h"
 #include "Engine/Graphics/Renderer/Renderer.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/Overlays.h"
