+++ src/Engine/Events/Processor.cpp
@@ -8,7 +8,7 @@
 #include "Engine/mm7_data.h"
 #include "Engine/Graphics/LocationFunctions.h"
 #include "Engine/Objects/DecorationList.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/Objects/Actor.h"
 #include "Engine/Events/EventMap.h"
