+++ src/Engine/Graphics/Viewport.cpp
@@ -3,7 +3,7 @@
 #include "Engine/Engine.h"
 #include "Engine/Events/Processor.h"
 #include "Engine/Objects/DecorationList.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/Vis.h"
