+++ src/Engine/Graphics/Vis.cpp
@@ -8,7 +8,7 @@
 #include "Engine/Engine.h"
 #include "Engine/OurMath.h"
 
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Graphics/BspRenderer.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Indoor.h"
