+++ src/Engine/Objects/CMakeLists.txt
@@ -4,6 +4,7 @@ set(ENGINE_OBJECTS_SOURCES
         Actor.cpp
         Chest.cpp
         CombinedSkillValue.cpp
+        Decoration.cpp
         DecorationList.cpp
         ItemEnumFunctions.cpp
         Items.cpp
@@ -21,6 +22,7 @@ set(ENGINE_OBJECTS_HEADERS
         Chest.h
         ChestEnums.h
         CombinedSkillValue.h
+        Decoration.h
         DecorationList.h
         DecorationEnums.h
         ItemEnchantment.h
