+++ src/Engine/Objects/Decoration.cpp
@@ -1,11 +1,11 @@
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Decoration.h"
 
 #include <cstdlib>
 #include <algorithm>
 #include <vector>
 
-#include "Engine/Party.h"
 #include "Engine/Random/Random.h"
+#include "Engine/Party.h"
 
 #include "Library/Logger/Logger.h"
