+++ src/Engine/Snapshots/EntitySnapshots.cpp
@@ -7,7 +7,7 @@
 #include "Engine/ArenaEnumFunctions.h"
 #include "Engine/Engine.h"
 #include "Engine/Graphics/Indoor.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Objects/DecorationList.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Overlays.h"
