+++ src/GUI/UI/Books/MapBook.cpp
@@ -6,7 +6,7 @@
 #include "Engine/Events/Processor.h"
 #include "Engine/Graphics/Renderer/Renderer.h"
 #include "Engine/Graphics/Indoor.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/Graphics/Image.h"
