+++ src/GUI/UI/UIBranchlessDialogue.cpp
@@ -3,7 +3,7 @@
 #include "Engine/AssetsManager.h"
 #include "Engine/Events/Processor.h"
 #include "Engine/Graphics/Renderer/Renderer.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Party.h"
 #include "Engine/mm7_data.h"
