+++ src/GUI/UI/UIGame.cpp
@@ -12,7 +12,7 @@
 #include "Engine/Events/Processor.h"
 #include "Engine/Graphics/BSPModel.h"
 #include "Engine/Objects/DecorationList.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/Renderer/Renderer.h"
