+++ src/Media/Audio/AudioPlayer.cpp
@@ -9,7 +9,7 @@
 #include <memory>
 
 #include "Engine/Graphics/Indoor.h"
-#include "Engine/Graphics/Level/Decoration.h"
+#include "Engine/Objects/Decoration.h"
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/Spells/Spells.h"
